Support pytest

+++ pyspike/cython/cython_directionality.pyx
@@ -35,7 +35,7 @@ from libc.math cimport fmin
 
 # from pyspike.cython.cython_distances cimport get_tau
 
-DTYPE = np.float
+DTYPE = float
 ctypedef np.float_t DTYPE_t
